Add files via upload
Debugged the random graph generator (moreover some logical errors, of which the one that checks whether a vertex was already adjacent to that vertex is a major one). One other note: we need to decide how we count vertices and edges - do we start at 1 or 0? We are mixing them up at the moment: arrays start at 0, in this class vertices are printed as 1 if we mean vertex 1, while our algorithms of the previous phase use 0 for vertex one - which is way more logical. But we need to make a final decision and we all need to keep track of this!!!