Alter the game's structure
The Game class now extends JFrame. The Level class has been changed, it's now called State and extends JPanel, thus the functionality of Swing can be used in there. Additionally, everytime a State is entered or exited, enter() and exit() gets invoked, respectively. The GameState class keeps track of the current state of the game, and is instantiated in the Game class. Also, a new class has been added: GameMode, this should serve as a basis of all game modes and extends State. It has Canvas which is located at BorderLayout.Center. Thus, a HUD implemented in Swing could be added to it. Invoking it's enter() method starts a thread, which has the functionality of the former main game loop and draws the graph on prior mentioned Canvas. Also event handling is implemented, thus all vertices are, by default, draggable and coloUrable (with the ColorSelectionMenu). So, for implementing a new game mode just extend the class and alter the functionality in the way you need.